feat:동의한 항목 데이터 받아오기
