+++ lib/viewmodels/auth/kakao_auth_service.dart
@@ -4,6 +4,7 @@ import 'package:crypto/crypto.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
 import 'package:nota_note/models/user_model.dart';
+import 'package:nota_note/viewmodels/auth/auth_common.dart';
 
 final kakaoAuthViewModelProvider = Provider((ref) => KakaoAuthViewmodel());
 
@@ -12,38 +13,43 @@ class KakaoAuthViewmodel {
 
   Future<void> signInWithKakao() async {
     try {
-      // 1. 카카오톡 앱 or 계정으로 로그인
+      // 카카오톡 앱 or 계정으로 로그인
       bool isInstalled = await isKakaoTalkInstalled();
       OAuthToken token = isInstalled
-          ? await UserApi.instance.loginWithKakaoTalk()
-          : await UserApi.instance.loginWithKakaoAccount();
+          ? await UserApi.instance.loginWithKakaoTalk() //카카오톡 앱(우선)
+          : await UserApi.instance.loginWithKakaoAccount(); //카카오 계정
 
-      // 2. 사용자 정보 조회
+      // 사용자 정보 조회
       final user = await UserApi.instance.me();
       final kakaoAccount = user.kakaoAccount;
       final userId = user.id.toString();
 
       print('카카오 로그인 성공: ${kakaoAccount?.email}');
 
-      // 3. UID 기반 해시태그 생성
-      String generateHashedTag(String uid) {
-        final bytes = utf8.encode(uid);
-        final digest = sha256.convert(bytes);
-        return '#${digest.toString().substring(0, 6)}';
-      }
+      // //  UID 기반 해시태그 생성
+      // String generateHashedTag(String uid) {
+      //   final bytes = utf8.encode(uid);
+      //   final digest = sha256.convert(bytes);
+      //   return '#${digest.toString().substring(0, 6)}';
+      // }
+
+      //  동의된 정보만 저장 (null-safe 처리)
+      final nickname = kakaoAccount?.profile?.nickname;
+      final profileImage = kakaoAccount?.profile?.profileImageUrl;
+      final email = kakaoAccount?.email;
 
       final userModel = UserModel(
         userId: userId,
-        displayName: kakaoAccount?.profile?.nickname ?? 'NoName',
-        email: kakaoAccount?.email ?? 'unknown@email.com',
-        photoUrl: kakaoAccount?.profile?.profileImageUrl ?? '',
+        displayName: nickname ?? 'NoName',
+        email: email ?? 'unknown@email.com',
+        photoUrl: profileImage ?? '',
         hashTag: generateHashedTag(userId),
         loginProviders: 'kakao',
-        createdAt: DateTime.now(),
+        createdAt: DateTime.now(), // 임시 표시, Firestore에서 서버시간으로 덮어씀
         updatedAt: DateTime.now(),
       );
 
-      // 4. Firestore에 저장
+      // Firestore에 저장
       await _firestore
           .collection('users')
           .doc(userId)
